refactor(portfolio): extract transforms into PortfolioTransformMixin
Move truncate/lag/smoothed_holding/correlation into
_portfolio_transform.py, matching the existing _portfolio_* mixin
pattern. New portfolios are built via type(self).from_cash_position
(returns Self, behaviour-identical to the prior Portfolio(...) calls).
